Citation Editor working on Mac.
Uses Carbon API to bring Mendeley or the word processor window to the foreground.

Positions Citation Editor in the middle of the screen, or where the user last positioned it (until restarting MD).

Search results not drawing properly. RichTextDelegate::sizeHint() doesn't return the correct size. I can't figure out why it returns a sensible size when used for the search results in Mendeley Desktop's DocumentsView but not in the Citation Editor's QListView.

refs #13363 (8)