Fix hang after crash on Windows XP with corrupt search index.
Following a segfault, the breakpad crash handler was invoked, which
saved a minidump file.  On Windows XP, ExitProcess() was then invoked
and a hang occurred during the DllMain() prelude when running a global
destructor in QtCore.   The hang does not occur under Windows 7 or if
the Windows Error Reporting dialog was enabled - where the process is terminated
when the user clicks the 'Close' button in the dialog.

This commit fixes the problem by calling TerminateProcess() immediately
after writing the minidump file, preventing any further user-mode code
from running in the process.  This has the same effect as ending the process
via the task manager.

Calling TerminateProcess() will prevent debuggers or other handlers doing anything
when the exception occurs by default.  This can be worked around by setting the
MENDELEY_DISABLE_BREAKPAD environment variable before running a release build
in the debugger (see http://stackoverflow.com/questions/378380/how-can-i-debug-a-win32-process-that-unexpectedly-terminates-silently
for other alternatives).

Fixes #17814 (7)
Reviewed-by: Nicolas Esteves