Fix passing of module names to external symbol fetchers on Windows
On Windows, CodeModule::debug_file() may return the full path to
the module from the system where the binary was generated.

On the symbol server however, only the basename of the file
is used. Strip the directory path from the module name and
escape the binary name (in case it contains spaces etc.) before
passing it to popen().

MD-20322