Fix failure to fetch debug symbols for modules whose names contain spaces
Requests to fetch debug symbol files for modules whose names contained
spaces (eg. 'Mendeley Desktop') failed because urllib2.Request() did
not escape the URL.

This resulted in the script sending an opening request with the line
'GET /.../Module Name.sym HTTP/1.1' with the part of the URL after
the space being interpreted as the HTTP marker. AWS then rejected the
request with an HTTP 505.

Escape the path part of the URL before passing it to urllib2.Request()

MD-20591