Support fetching symbols under alternative debug file names than the one specified in the minidump
The minidump contains the full path to each module's binary, minidump_stackwalk
strips the path and then invokes fetch-symbols.py with the result to fetch
debug symbols.

This breaks if the debug symbols are originally extracted from the binary with one name
and the binary has a different name when it is deployed on the end user's system.

This happens for Ubuntu builds of Mendeley Desktop where the main binary is named
mendeleydesktop.i486 or mendeleydesktop.x86_64 on the build machine but
installed as /opt/mendeleydesktop/bin/mendeleydesktop

The workaround is to support an option in fetch-symbols.py to read a list
of alternative names for a binary from a JSON config file.